Helm 2.16.6

+++ assets/common.sh
@@ -4,51 +4,65 @@ set -e
 setup_kubernetes() {
   payload=$1
   source=$2
-  # Setup kubectl
-  cluster_url=$(jq -r '.source.cluster_url // ""' < $payload)
-  if [ -z "$cluster_url" ]; then
-    echo "invalid payload (missing cluster_url)"
-    exit 1
-  fi
-  if [[ "$cluster_url" =~ https.* ]]; then
 
-    cluster_ca=$(jq -r '.source.cluster_ca // ""' < $payload)
-    admin_key=$(jq -r '.source.admin_key // ""' < $payload)
-    admin_cert=$(jq -r '.source.admin_cert // ""' < $payload)
-    token=$(jq -r '.source.token // ""' < $payload)
-    token_path=$(jq -r '.params.token_path // ""' < $payload)
+  mkdir -p /root/.kube
+  kubeconfig_path=$(jq -r '.params.kubeconfig_path // ""' < $payload)
+  absolute_kubeconfig_path="${source}/${kubeconfig_path}"
+  if [ -f "$absolute_kubeconfig_path" ]; then
+    cp "$absolute_kubeconfig_path" "/root/.kube/config"
+  else
+    # Setup kubectl
+    cluster_url=$(jq -r '.source.cluster_url // ""' < $payload)
+    if [ -z "$cluster_url" ]; then
+      echo "invalid payload (missing cluster_url)"
+      exit 1
+    fi
+    if [[ "$cluster_url" =~ https.* ]]; then
 
-    mkdir -p /root/.kube
+      cluster_ca=$(jq -r '.source.cluster_ca // ""' < $payload)
+      admin_key=$(jq -r '.source.admin_key // ""' < $payload)
+      admin_cert=$(jq -r '.source.admin_cert // ""' < $payload)
+      token=$(jq -r '.source.token // ""' < $payload)
+      token_path=$(jq -r '.params.token_path // ""' < $payload)
 
-    ca_path="/root/.kube/ca.pem"
-    echo "$cluster_ca" | base64 -d > $ca_path
-    kubectl config set-cluster default --server=$cluster_url --certificate-authority=$ca_path
+      ca_path="/root/.kube/ca.pem"
+      echo "$cluster_ca" | base64 -d > $ca_path
+      kubectl config set-cluster default --server=$cluster_url --certificate-authority=$ca_path
 
-    if [ -f "$source/$token_path" ]; then
-      kubectl config set-credentials admin --token=$(cat $source/$token_path)
-    elif [ ! -z "$token" ]; then
-      kubectl config set-credentials admin --token=$token
+      if [ -f "$source/$token_path" ]; then
+        kubectl config set-credentials admin --token=$(cat $source/$token_path)
+      elif [ ! -z "$token" ]; then
+        kubectl config set-credentials admin --token=$token
+      else
+        mkdir -p /root/.kube
+        key_path="/root/.kube/key.pem"
+        cert_path="/root/.kube/cert.pem"
+        echo "$admin_key" | base64 -d > $key_path
+        echo "$admin_cert" | base64 -d > $cert_path
+        kubectl config set-credentials admin --client-certificate=$cert_path --client-key=$key_path
+      fi
+
+      kubectl config set-context default --cluster=default --user=admin
     else
-      key_path="/root/.kube/key.pem"
-      cert_path="/root/.kube/cert.pem"
-      echo "$admin_key" | base64 -d > $key_path
-      echo "$admin_cert" | base64 -d > $cert_path
-      kubectl config set-credentials admin --client-certificate=$cert_path --client-key=$key_path
+      kubectl config set-cluster default --server=$cluster_url
+      kubectl config set-context default --cluster=default
     fi
 
-    kubectl config set-context default --cluster=default --user=admin
-  else
-    kubectl config set-cluster default --server=$cluster_url
-    kubectl config set-context default --cluster=default
+    kubectl config use-context default
   fi
 
-  kubectl config use-context default
   kubectl version
 }
 
 setup_tls() {
   tls_enabled=$(jq -r '.source.tls_enabled // "false"' < $payload)
+  tillerless=$(jq -r '.source.tillerless // "false"' < $payload)
   if [ "$tls_enabled" = true ]; then
+    if [ "$tillerless" = true ]; then
+      echo "Setting both tls_enabled and tillerless is not supported"
+      exit 1
+    fi
+
     helm_ca=$(jq -r '.source.helm_ca // ""' < $payload)
     helm_key=$(jq -r '.source.helm_key // ""' < $payload)
     helm_cert=$(jq -r '.source.helm_cert // ""' < $payload)
@@ -74,11 +88,57 @@ setup_tls() {
 }
 
 setup_helm() {
+  # $1 is the name of the payload file
+  # $2 is the name of the source directory
   init_server=$(jq -r '.source.helm_init_server // "false"' < $1)
-  tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+
+  # Compute tiller_namespace as follows:
+  # If kubeconfig_tiller_namespace is set, then tiller_namespace is the namespace from the kubeconfig
+  # If tiller_namespace is set and it is the name of a file, then tiller_namespace is the contents of the file
+  # If tiller_namespace is set and it is not the name of a file, then tiller_namespace is the literal
+  # Otherwise tiller_namespace defaults to kube-system
+  kubeconfig_tiller_namespace=$(jq -r '.source.kubeconfig_tiller_namespace // "false"' <$1)
+  if [ "$kubeconfig_tiller_namespace" = "true" ]
+  then
+    tiller_namespace=$(kubectl config view --minify -ojson | jq -r .contexts[].context.namespace)
+  else
+    tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+    if [ "$tiller_namespace" != "kube-system" -a -f "$2/$tiller_namespace" ]
+    then
+      tiller_namespace=$(cat "$2/$tiller_namespace")
+    fi
+  fi
+
+  tillerless=$(jq -r '.source.tillerless // "false"' < $payload)
   tls_enabled=$(jq -r '.source.tls_enabled // "false"' < $payload)
+  history_max=$(jq -r '.source.helm_history_max // "0"' < $1)
+  stable_repo=$(jq -r '.source.stable_repo // ""' < $payload)
+
+  if [ "$tillerless" = true ]; then
+    echo "Using tillerless helm"
+    helm_bin="helm tiller run ${tiller_namespace} -- helm"
+  else
+    helm_bin="helm"
+  fi
+
+  if [ -n "$stable_repo" ]; then
+    echo "Stable Repo URL : ${stable_repo}"
+    stable_repo="--stable-repo-url=${stable_repo}"
+  fi
+
   if [ "$init_server" = true ]; then
+    if [ "$tillerless" = true ]; then
+      echo "Setting both init_server and tillerless is not supported"
+      exit 1
+    fi
     tiller_service_account=$(jq -r '.source.tiller_service_account // "default"' < $1)
+
+    helm_init_wait=$(jq -r '.source.helm_init_wait // "false"' <$1)
+    helm_init_wait_arg=""
+    if [ "$helm_init_wait" = "true" ]; then
+      helm_init_wait_arg="--wait"
+    fi
+
     if [ "$tls_enabled" = true ]; then
       tiller_key=$(jq -r '.source.tiller_key // ""' < $payload)
       tiller_cert=$(jq -r '.source.tiller_cert // ""' < $payload)
@@ -95,19 +155,29 @@ setup_helm() {
       helm_ca_cert_path="/root/.helm/ca.pem"
       echo "$tiller_key" > $tiller_key_path
       echo "$tiller_cert" > $tiller_cert_path
-      helm tiller run $tiller_namespace -- helm init --tiller-tls --tiller-tls-cert $tiller_cert_path --tiller-tls-key $tiller_key_path --tiller-tls-verify --tls-ca-cert $tiller_key_path --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --upgrade
+      helm tiller run $tiller_namespace -- helm init --tiller-tls --tiller-tls-cert $tiller_cert_path --tiller-tls-key $tiller_key_path --tiller-tls-verify --tls-ca-cert $tiller_key_path --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --upgrade $helm_init_wait_arg
     else
-      helm tiller run $tiller_namespace -- helm init --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --upgrade
+      helm tiller run $tiller_namespace -- helm init --tiller-namespace=$tiller_namespace --service-account=$tiller_service_account --upgrade $helm_init_wait_arg
     fi
     wait_for_service_up tiller-deploy 10
   else
     export HELM_HOST=$(jq -r '.source.helm_host // ""' < $1)
     helm tiller run $tiller_namespace -- helm init -c --tiller-namespace $tiller_namespace > /dev/null
   fi
+
+  tls_enabled_arg=""
   if [ "$tls_enabled" = true ]; then
-    helm tiller run $tiller_namespace -- helm version --tls --tiller-namespace $tiller_namespace
-  else
-    helm tiller run $tiller_namespace -- helm version --tiller-namespace $tiller_namespace
+    tls_enabled_arg="--tls"
+  fi
+  helm tiller run $tiller_namespace -- helm $tls_enabled_arg --tiller-namespace $tiller_namespace
+
+  helm_setup_purge_all=$(jq -r '.source.helm_setup_purge_all // "false"' <$1)
+  if [ "$helm_setup_purge_all" = "true" ]; then
+    local release
+    for release in $(helm tiller run $tiller_namespace -- helm ls -aq --tiller-namespace $tiller_namespace )
+    do
+      helm tiller run $tiller_namespace -- helm delete $tls_enabled_arg --purge "$release" --tiller-namespace $tiller_namespace
+    done
   fi
 }
 
@@ -127,9 +197,27 @@ wait_for_service_up() {
 
 setup_repos() {
   repos=$(jq -c '(try .source.repos[] catch [][])' < $1)
-  tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+  plugins=$(jq -c '(try .source.plugins[] catch [][])' < $1)
+
+  kubeconfig_tiller_namespace=$(jq -r '.source.kubeconfig_tiller_namespace // "false"' <$1)
+  if [ "$kubeconfig_tiller_namespace" = "true" ]
+  then
+    tiller_namespace=$(kubectl config view --minify -ojson | jq -r .contexts[].context.namespace)
+  else
+    tiller_namespace=$(jq -r '.source.tiller_namespace // "kube-system"' < $1)
+  fi
 
   local IFS=$'\n'
+
+  for pl in $plugins; do
+    plurl=$(echo $pl | jq -cr '.url')
+    plversion=$(echo $pl | jq -cr '.version // ""')
+    if [ -n "$plversion" ]; then
+      plversionflag="--version $plversion"
+    fi
+    helm plugin install $plurl $plversionflag
+  done
+
   for r in $repos; do
     name=$(echo $r | jq -r '.name')
     url=$(echo $r | jq -r '.url')
@@ -148,10 +236,15 @@ setup_repos() {
 }
 
 setup_resource() {
+  tracing_enabled=$(jq -r '.source.tracing_enabled // "false"' < $1)
+  if [ "$tracing_enabled" = "true" ]; then
+    set -x
+  fi
+
   echo "Initializing kubectl..."
   setup_kubernetes $1 $2
   echo "Initializing helm..."
   setup_tls $1
-  setup_helm $1
+  setup_helm $1 $2
   setup_repos $1
 }
